requirements.txt
aiohttp==3.9.5
aiohttp-retry==2.8.3
aiosignal==1.3.1
amqp==5.2.0
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
appdirs==1.4.4
asyncssh==2.14.2
atpublic==4.1.0
attrs==23.2.0
billiard==4.2.0
celery==5.4.0
certifi==2024.6.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
configobj==5.0.8
cryptography==42.0.8
dictdiffer==0.9.0
diskcache==5.6.3
distro==1.9.0
dpath==2.2.0
dulwich==0.22.1
dvc==3.51.2
dvc-data==3.15.1
dvc-http==2.32.0
dvc-objects==5.1.0
dvc-render==1.0.2
dvc-studio-client==0.20.0
dvc-task==0.4.0
entrypoints==0.4
filelock==3.15.1
flatten-dict==0.4.2
flufl.lock==7.1.1
frozenlist==1.4.1
fsspec==2024.6.0
funcy==2.0
gitdb==4.0.11
GitPython==3.1.43
grandalf==0.8
gto==1.7.1
hydra-core==1.3.2
idna==3.7
iterative-telemetry==0.0.8
joblib==1.4.2
kombu==5.3.7
markdown-it-py==3.0.0
mdurl==0.1.2
multidict==6.0.5
networkx==3.3
nltk==3.8.1
numpy==2.0.0
omegaconf==2.3.0
orjson==3.10.5
packaging==24.1
pandas==2.2.2
pathspec==0.12.1
platformdirs==3.11.0
prompt_toolkit==3.0.47
psutil==6.0.0
pycparser==2.22
pydantic==2.7.4
pydantic_core==2.18.4
pydot==2.0.0
pygit2==1.15.0
Pygments==2.18.0
pygtrie==2.5.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
pytz==2024.1
pywin32==306
PyYAML==6.0.1
regex==2024.5.15
requests==2.32.3
rich==13.7.1
ruamel.yaml==0.18.6
ruamel.yaml.clib==0.2.8
scikit-learn==1.5.0
scipy==1.13.1
scmrepo==3.3.5
semver==3.0.2
shellingham==1.5.4
shortuuid==1.0.13
shtab==1.7.1
six==1.16.0
smmap==5.0.1
sqltrie==0.11.0
tabulate==0.9.0
threadpoolctl==3.5.0
tomlkit==0.12.5
tqdm==4.66.4
typer==0.12.3
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
vine==5.1.0
voluptuous==0.14.2
wcwidth==0.2.13
yarl==1.9.4
zc.lockfile==3.0.post1
-e .